move common vertx sql code
