--- dd-java-agent/instrumentation/vertx-mysql-client-4.0/src/main/java/datadog/trace/instrumentation/vertx_sql_client_4/CursorImplInstrumentation.java
@@ -1,47 +1,0 @@
-package datadog.trace.instrumentation.vertx_sql_client_4;
-
-import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-
-import com.google.auto.service.AutoService;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.agent.tooling.InstrumenterModule;
-import java.util.Map;
-
-@AutoService(InstrumenterModule.class)
-public class CursorImplInstrumentation extends InstrumenterModule.Tracing
-    implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
-  public CursorImplInstrumentation() {
-    super("vertx", "vertx-sql-client");
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    return singletonMap("io.vertx.sqlclient.PreparedStatement", "datadog.trace.api.Pair");
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      packageName + ".QueryResultHandlerWrapper", packageName + ".VertxSqlClientDecorator",
-    };
-  }
-
-  @Override
-  public String instrumentedType() {
-    return "io.vertx.sqlclient.impl.CursorImpl";
-  }
-
-  @Override
-  public void methodAdvice(MethodTransformer transformer) {
-    transformer.applyAdvice(
-        isMethod()
-            .and(isPublic())
-            .and(named("read"))
-            .and(takesArgument(1, named("io.vertx.core.Handler"))),
-        packageName + ".CursorReadAdvice");
-  }
-}
