--- dd-java-agent/instrumentation/vertx-pg-client-4.2.0/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4_2_0/PrepareHandlerWrapper.java
@@ -1,32 +1,0 @@
-package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
-
-import datadog.trace.api.Pair;
-import datadog.trace.bootstrap.ContextStore;
-import datadog.trace.bootstrap.instrumentation.jdbc.DBInfo;
-import datadog.trace.bootstrap.instrumentation.jdbc.DBQueryInfo;
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Handler;
-import io.vertx.sqlclient.PreparedStatement;
-
-public class PrepareHandlerWrapper implements Handler<AsyncResult<PreparedStatement>> {
-  private final Handler<AsyncResult<PreparedStatement>> handler;
-  private final ContextStore<PreparedStatement, Pair> contextStore;
-  private final Pair<DBInfo, DBQueryInfo> queryInfo;
-
-  public PrepareHandlerWrapper(
-      Handler<AsyncResult<PreparedStatement>> handler,
-      ContextStore<PreparedStatement, Pair> contextStore,
-      Pair<DBInfo, DBQueryInfo> queryInfo) {
-    this.handler = handler;
-    this.contextStore = contextStore;
-    this.queryInfo = queryInfo;
-  }
-
-  @Override
-  public void handle(AsyncResult<PreparedStatement> event) {
-    if (event.succeeded()) {
-      contextStore.put(event.result(), queryInfo);
-    }
-    handler.handle(event);
-  }
-}
