+++ dd-java-agent/instrumentation/vertx-sql-common/src/main/java/datadog/trace/instrumentation/vertx_sql_common/QueryAdvice.java
@@ -1,9 +1,9 @@
-package datadog.trace.instrumentation.vertx_sql_client_4;
+package datadog.trace.instrumentation.vertx_sql_common;
 
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activeSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.captureSpan;
-import static datadog.trace.instrumentation.vertx_sql_client_4.VertxSqlClientDecorator.DECORATE;
+import static datadog.trace.instrumentation.vertx_sql_common.VertxSqlClientDecorator.DECORATE;
 
 import datadog.trace.api.Pair;
 import datadog.trace.bootstrap.ContextStore;
@@ -12,7 +12,6 @@ import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
-import io.vertx.mysqlclient.MySQLConnection;
 import io.vertx.sqlclient.Query;
 import io.vertx.sqlclient.SqlResult;
 import net.bytebuddy.asm.Advice;
@@ -25,11 +24,6 @@ public class QueryAdvice {
         @Advice.This final Query<?> zis, @Advice.Return final Query<?> ret) {
       ContextStore<Query, Pair> contextStore = InstrumentationContext.get(Query.class, Pair.class);
       contextStore.put(ret, contextStore.get(zis));
-    }
-
-    // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
-    private static void muzzleCheck(MySQLConnection connection) {
-      connection.ping();
     }
   }
 
@@ -72,11 +66,6 @@ public class QueryAdvice {
       if (null != clientScope) {
         clientScope.close();
       }
-    }
-
-    // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
-    private static void muzzleCheck(MySQLConnection connection) {
-      connection.ping();
     }
   }
 }
