+++ dd-java-agent/instrumentation/vertx-sql-common/src/main/java/datadog/trace/instrumentation/vertx_sql_common/SqlConnectionBasePrepareAdvice.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
+package datadog.trace.instrumentation.vertx_sql_common;
 
 import datadog.trace.api.Pair;
 import datadog.trace.bootstrap.InstrumentationContext;
@@ -6,6 +6,7 @@ import datadog.trace.bootstrap.instrumentation.jdbc.DBInfo;
 import datadog.trace.bootstrap.instrumentation.jdbc.DBQueryInfo;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
+import io.vertx.mysqlclient.MySQLConnection;
 import io.vertx.sqlclient.PreparedStatement;
 import io.vertx.sqlclient.SqlClient;
 import net.bytebuddy.asm.Advice;
@@ -25,5 +26,10 @@ public class SqlConnectionBasePrepareAdvice {
     handler =
         new PrepareHandlerWrapper(
             handler, InstrumentationContext.get(PreparedStatement.class, Pair.class), info);
+  }
+
+  // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
+  private static void muzzleCheck(MySQLConnection connection) {
+    connection.ping();
   }
 }
