+++ settings.gradle
@@ -506,6 +506,7 @@ include ':dd-java-agent:instrumentation:vertx-pg-client-4.2.0'
 include ':dd-java-agent:instrumentation:vertx-redis-client-3.9'
 include ':dd-java-agent:instrumentation:vertx-redis-client-3.9:stubs'
 include ':dd-java-agent:instrumentation:vertx-rx-3.5'
+include ':dd-java-agent:instrumentation:vertx-sql-common'
 include ':dd-java-agent:instrumentation:vertx-web-3.4'
 include ':dd-java-agent:instrumentation:vertx-web-3.5'
 include ':dd-java-agent:instrumentation:vertx-web-3.9'
